Mon, Jan 1, 2024, 10:23 PM -06:00

+++ Week 5/test_fuel/fuel.py
@@ -8,13 +8,13 @@ def convert(fraction):
         try:
             x,y = fraction.split("/")
             percentage = int(x) / int(y) * 100
-            percentage = round(percentage)
+            percentage = int(round(percentage))
         except (ValueError, ZeroDivisionError):
             continue
         if percentage <0 or percentage >100:
             continue
         else:
-            return int(percentage)
+            return percentage
 
 def gauge(percentage):
     if percentage <= 1:
